add SingleSingletonStore
