+++ mex/common/context.py
@@ -32,3 +32,36 @@ class SingletonStore(Generic[_T]):
     def reset(self) -> None:
         """Remove all singleton instances from the store."""
         self._instances_by_class.clear()
+
+
+class SingleSingletonStore(Generic[_T]):
+    """Thin wrapper for storing a single thread-local singleton.
+
+    Stores only a single instance. Requested class must either be
+    the same or a parent class of the stored class.
+    """
+
+    def __init__(self) -> None:
+        """Create a new settings singleton store."""
+        self._singleton: _T | None = None
+
+    def load(self, cls: type[_T]) -> _T:
+        """Retrieve the settings for the given class or create a new one."""
+        if self._singleton is None:
+            self._singleton = cls()
+            return self._singleton
+        if not issubclass(type(self._singleton), cls):
+            raise TypeError(
+                f"requested class ({cls}) is not not a parent class of loaded class "
+                f"({type(self._singleton)}). "
+                f"Did you initialize {cls} upon startup?"
+            )
+        return self._singleton
+
+    def push(self, instance: _T) -> None:
+        """Set or replace a singleton instance in the store."""
+        self._singleton = instance
+
+    def reset(self) -> None:
+        """Remove singleton instance from the store."""
+        self._singleton = None
